Add content access comp & cookie handler
First pass at setting and using of contentAccess cookie handler & new content access compoents.

At the moment the with content contentFormater is handeling the generation of the content access state which includes the cookie details.  @todo figure out if this is the best place to set this while relying on the req within the content template to determine if it should display the gate.  Mainly because this is base on the req.cookie being read.